backend/requirements.txt
asgiref
Django
django-cors-headers
django-extensions
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
pillow
sqlparse
tzdata
psycopg2-binary
python-dotenv